pyproject.toml
[tool.poetry]
name = "roo"
version = "0.11.0"
description = "A package manager to handle R environments"
authors = ["Stefano Borini <stefano.borini@astrazeneca.com>"]
license = "Apache-2.0"
packages = [
    { include = "roo", from = "src" },
]

[tool.poetry.scripts]
roo = 'roo.cli.__main__:main'

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22.0"
click = "^7.0"
beautifulsoup4 = "^4.8.2"
toml = "^0.10.0"
packaging = "^20.1"
atomicwrites = "^1.4"
rich = "^2.2.3"
GitPython = "^3.1.13"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
flake8 = "^3.7.9"
pre-commit = "^2.5.1"
autopep8 = "^1.5.3"
mypy = "0.800"
sphinx = "^3.3.1"
pydata_sphinx_theme = "^0.4.1"
types-toml = "^0.1.1"
bumpversion = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
